+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 # UNRELEASED
 
+* M365DSCGraphShim
+  * Added handling for PowerShell 7.5 with Mgx to improve Graph calls.
+* DEPENDENCIES
+  * Added `Mgx` with version `1.0.2`.
+* MISC
+  * Invoke PowerShell 7 for all resources if not already running under it.
+
+# UNRELEASED
+
 * AADGroup
   * Fixed an issue where `GroupTypes` would be returned as a null object.
     FIXES [#7253](https://github.com/Microsoft365DSC/Microsoft365DSC/issues/7253)